+++ src/index.js
@@ -1,0 +1,3 @@
+import counter from "./card-counter/card-counter-service";
+
+counter.count();