+++ src/runtime/background.js
@@ -1,5 +1,5 @@
 chrome.runtime.onInstalled.addListener(function () {
 
-    console.log('The color is green.');
+    console.log('Trello Toolkit is active!');
 
 });