+++ CHANGELOG.md
@@ -16,3 +16,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ### Added
 
+-  Added Card Counter feature, showing the number of cards on each Trello List Header.